Moved the MediaSourceImageCard and MediaSourceImageCardEditor classes to their own files. Updated some instructions in the README.md
